requirement.txt
# torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121
easy-torch==1.3.2
easydict==1.10
setproctitle
# wandb
numpy==1.23.0
pandas==1.5.1
einops
scipy
packaging
scikit-learn
tables==3.7.0